add groups search
